Modiicaciones pertinenentes de archivos importados en main
