--- 18TE0284_JoseGpeCruzValera/assets.py
@@ -1,20 +1,0 @@
-import time
-
-def cargando():
-    print('Cargando...')
-    time.sleep(0.5)
-    print('...')
-    time.sleep(0.5)
-    print('...')
-    time.sleep(0.5)
-    print(' 1 nueva notificacion')
-
-def cargando2():
-    print("\nGracias por usar el programa")
-    print('Cerrando programa ⏱️')
-    print("\n3..")
-    time.sleep(0.5)
-    print("   2..")
-    time.sleep(0.75)
-    print("       1..")
-    time.sleep(0.75)
