+++ 18TE0284_JoseGpeCruzValera/main.py
@@ -1,29 +1,37 @@
 import Gerente as Ceo
-import empleado as emp
+import Empleado as emp
 import Vestimenta as Clothes
 import time
-import assets
+import Assetstools as assets
 
 
-
+#creamos el objeto CEO -------------------------------------------------
 CEOdEmpresar = Ceo.Gerente('Jose Antonio Ponton',45,39999.99,str(85),1.80)
 pantalonGerente = Clothes.pantalon("Gris",'Mediana 34','LV','Hombre')
 zapatosGerente = Clothes.zapatos("Negro","27","Flexi","Masculino")
 sacoGerente = Clothes.saco('Azul','Mediana 36','Rey Palomo','Hombre','Varias telas finas',3)
 camisetaGerente = Clothes.camiseta('azul','Mediana 34','Lacoste')
+#creamos su ropa de trabajo -------------------------------------------------
+#       Pantalon
+#           Camisa
+#               Zapatos
+#                   Saco con los 3 botones
 
+#crearemos el obejeto Secretaria---------------------------------------------------------
 Secretaria = Ceo.persona('Maria Buena Aventura',25,1500.00,'b1',1.65)
-print(' Hola buenas tardes mi nombre es {} le acabo de enviar un correo de la empresa para su solicitud \
-    de empleo, en breves momentos le llegara...'.format(Secretaria.get_name()))
+print(""" Hola buenas tardes mi nombre es {} le acabo de enviar un correo de la empresa para su 
+    solicitud de empleo, en breves momentos le llegara...""".format(Secretaria.get_name()))
+#----------------------------------------------------------------------------------------------------------------
+
 assets.cargando()
 
 CEOdEmpresar.get_invitacion()
 
-
+print('Bienvenido al programa de Entrevista de trabajo')
 
 def main():
     
-    print("""\nBienvenido al programa de Entrevista de trabajo\n
+    print("""\n\n
     ¿Qué desea hacer?
     [1] conocer el jefe
     [2] verte al espejo
@@ -45,7 +53,6 @@ def main():
         # CEO.verVestimenta()
         time.sleep(3)
         return main()
-    
         #-----------------------------------------------------s
     elif opcion == 2:
         pass
